Clean sort menu code
