+++ libs/@hashintel/ds-components/src/components/Menu/SelectableList/selectable-list.stories.tsx
@@ -87,16 +87,6 @@ export const Default: Story<SelectableListProps> = (args) => (
   </StaticMenu>
 );
 
-/**
- * Custom rows render arbitrary content on their own line. They are never
- * highlighted by arrow keys or mouse hover, but focusable children (the
- * input and the buttons) join keyboard navigation: Tab / Shift+Tab move
- * between whole blocks of regular items (entering a block highlights its
- * first — or, backwards, last — item) and custom-row focusables, and leave
- * the list past either end (to the surrounding buttons here). Arrow keys
- * pressed inside a custom row resume item navigation from that row's
- * position in the list.
- */
 export const CustomItems: Story<SelectableListProps> = (args) => (
   <div
     className={css({
