+++ libs/@hashintel/ds-components/src/main.ts
@@ -18,13 +18,14 @@ export type {
   Input as FilterInput,
   FilterValue,
 } from "./components/Filter/filter-util";
+export { SortMenu } from "./components/Filter/sort-menu";
 export {
+  readSavedSort,
   type SortDirection,
   type SortDirectionsAvailable,
   type Sorter,
-  SortMenu,
-} from "./components/Filter/sort-menu";
-export { readSavedSort, writeSavedSort } from "./components/Filter/sort-util";
+  writeSavedSort,
+} from "./components/Filter/sort-menu-util";
 export { Form } from "./components/Form/form";
 export { HelpTooltip } from "./components/HelpTooltip/help-tooltip";
 export { Icon, type IconName, iconNames } from "./components/Icon/icon";
